app/requirements.txt
llama-index==0.14.5
llama-index-cli==0.5.3
llama-index-core==0.14.5
llama-index-embeddings-huggingface==0.6.1
llama-index-embeddings-openai==0.5.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.2
llama-index-llms-openai==0.6.5
llama-index-readers-file==0.5.4
llama-index-readers-llama-parse==0.5.1
llama-index-vector-stores-postgres==0.7.0
llama-index-workflows==2.8.3
openai==1.109.1
pgvector==0.4.1
psycopg2-binary==2.9.11
python-dotenv==1.1.1
llama-index-llms-ollama==0.8.0
llama-index-vector-stores-postgres==0.7.0
